Move _some_ of util/images back to venia-ui because of dependency on venia-drivers
