+++ packages/venia-ui/lib/components/Gallery/item.js
@@ -4,10 +4,8 @@ import { Link, resourceUrl } from '@magento/venia-drivers';
 import { Price } from '@magento/peregrine';
 
 import classify from '../../classify';
-import {
-    transparentPlaceholder,
-    generateSrcset
-} from '@magento/peregrine/lib/util/images';
+import { transparentPlaceholder } from '@magento/peregrine/lib/util/images';
+import { generateSrcset } from '../../util/images';
 import defaultClasses from './item.css';
 
 // The placeholder image is 4:5, so we should make sure to size our product
